Added incremental vertical line drawing
It draws the lines in 40px segments because when i have a maze
data structure, i can lookup the current line being drawn and
choose to not draw it if it is a hole in the wall.

Due to nexted loops this current method is O(n^2) possibly
replace this in the future with a single loop and modular
arithmatic to change horizontal position.