Added the abity to draw edges of the maze
When the algorithm processes a cell it checks if it's on an edge of the maze.
If so it draws the maze edge.

At the moment i am "Drawing" the maze edges by ading their co-ordinates to
an array of wall positions (seperate ones for H or V).
In the future i will likely traverse these arrays of points and draw them
instead of using booleans and if statements, this will give simpler code.