Refactoring of results printer
Extracted the concept of an itterating for loop into a function called
forEach in the top level, this will be useful for much more than
logging errors.

I moved the celebration statement into a constant that is printed in
printCelebration to reduce cognative load.

I also fixed a test i broke to test the printer in the last commit.