Removed current maze drawer as it will not be used
The current maze drawin system was designed in about 1 hour,
and it was made before i had decided how to generate the maze.

The idea was that i would have an array of true and false values,
true if there was a wall there, false if there wasn't.
I would then have tried to draw every possible wall but first
check if it was supposed to be there from the array.
This seems inefficient in light of my new array of wall positions
And the new system to replace it would be as simple as itterating
throught the points and drawing a wall at each.