Various changes, see body
Deleted all code pertaining to drawing the mazes, as the system is old and
incompatible with the way i am currently generating the maze.
If you require the maze code, retreve it from a previous commit.

Added an add function to points that returns a new point containing the added
values of the calee and arguments. It takes either 1 point argument, or 2
ints for the addition x and y.

Am now using the times function to draw the starting maze blank template,
avoiding the repetition of simple for loops.

I am currently using a loop to process all of the top cells in the maze.
This will later develop into processing all cells for all types of walls, but
this gives me a way to start testing it rather than developing the whole thing
in one bang with no tests.

Extracted maze edge wall drawing to an object called MazeEndgeDrawer, i pass in
the addWall functions from maze as i need access to the maze data somehow but
don't want it housed in the MazeEdgeDrawer, this is probably a smell to be
fixed.

I am now alerting a report of a maze in order to see the changes to the maze
algorithm in real time without having to develop the new drawing system.

Extracted assertCreatesWall from the two specific wall type assertions

Started testing the new walls always generated at the top of the maze, much
more testing needed.